chore: test fixes
