+++ src/app/feature/theme/services/theme.service.mock.spec.ts
@@ -1,0 +1,11 @@
+import { ThemeService } from "./theme.service";
+
+export class MockThemeService implements Partial<ThemeService> {
+  ready() {
+    return true;
+  }
+  setTheme() {}
+  getCurrentTheme() {
+    return "mock_theme";
+  }
+}
