+++ src/app/shared/components/template/services/template-calc-functions/plh-calc-functions.spec.ts
@@ -38,6 +38,9 @@ describe("Template Calc - PLH Functions", () => {
 });
 
 describe("Template Calc - PLH Functions QA", () => {
+  beforeEach(() => {
+    (window as any).calc = PLH_CALC_FUNCTIONS;
+  });
   it("Handles string input", () => {
     const stringInput = JSON.stringify(MOCK_FAMILES());
     const res = PLH_CALC_FUNCTIONS.plh_add_family(stringInput, "Friedrich", "Graham", "Hannah");
