+++ src/app/shared/components/template/services/template-calc-functions/plh-calc-functions.ts
@@ -1,4 +1,4 @@
-import { IFunctionHashmap } from "packages/shared/src";
+import type { IFunctionHashmap } from "packages/shared/src/models/jsEvaluator/jsEvaluator";
 
 /**
  * Temporary functions used for plh sheets
